Create gw2api package
This package has the responsiblity of knowing all the detailed
information and data supplied by the gw2 api. This data is represented
in JSON format, and as such the gw2api package should house the go
structs required to parse the JSON data into. It also has the
responsibility of make the http requests to fetch the data